move register functions to utils
